:bricks: move modal into dedicated directory
